requirements.txt
PyYAML>=6.0
mp-api>=0.45.0
pymatgen>=2024.0.0
